+++ src/Application/_Env/Interfaces.ts
@@ -1,3 +1,5 @@
+import type { IStore } from 'Application/_Request/IStore';
+
 /**
  * Описание обобщенного window.location.
  * Выписаны те поля, которые есть на сервисе представления и в браузере
@@ -47,3 +49,47 @@ export interface ILocation {
  * @name Application/Env/ILocation#hash
  * @cfg {String} Возвращает #-часть URL
  */
+
+/**
+ * Набор опций для cookie
+ * @typedef {Object} ICookieOptions
+ * @property {String} domain domain
+ * @property {Number|Date} expires expires
+ * @property {String} path path
+ * @property {String} secure secure
+ */
+export interface ICookieOptions {
+    domain: string;
+    expires: number | Date;
+    path: string;
+    secure: string;
+}
+
+/**
+ * Интерфейс для работы с cookie
+ * @interface Application/Env/ICookie
+ * @extends Application/_Request/IStore
+ * @public
+ * @author Санников К.А.
+ */
+export interface ICookie extends IStore {
+    /**
+     * Получение значение из cookie
+     * @param {String}
+     */
+    get(key: string): string;
+    /**
+     * Устанавливаем cookie
+     * @param {String} key
+     * @param {String} value
+     * @param {Partial<ICookieOptions>} options
+     * @throws {Error} ошибка установки значения
+     */
+    set(key: string, value: string, options?: Partial<ICookieOptions>): boolean;
+    /**
+     * Удаляем cookie
+     * @param {String} key
+     * @throws {Error} ошибка очистки значения
+     */
+    remove(key: string): void;
+}
