+++ src/Application/_Interface/IEnv.ts
@@ -1,8 +1,7 @@
 /// <amd-module name="Application/_Interface/IEnv" />
 import type { IConsole } from 'Application/_Env/IConsole';
-import type { ILocation } from 'Application/_Env/Interfaces;
+import type { ICookie, ILocation } from 'Application/_Env/Interfaces';
 import { IConfig } from 'Application/_Interface/IConfig';
-import { ICookie } from 'Application/_Interface/ICookie';
 import { IHttpRequest } from 'Application/_Interface/IHttpRequest';
 import { IHttpResponse } from 'Application/_Interface/IHttpResponse';
 import { IRequest, IRequestInternal } from 'Application/_Interface/IRequest';
@@ -29,7 +28,7 @@ export interface IEnv {
  */
 /**
  * @name Application/_Interface/IEnv#cookie
- * @cfg {Application/_Interface/ICookie} cookie
+ * @cfg {Application/Env/ICookie} cookie
  */
 /**
  * @name Application/Env/IEnv#location
